+++ src/RenderEngine/GraphicsInstance.hpp
@@ -15,14 +15,15 @@ class GraphicsInstance {
 
 public:
   static vkb::Instance instance;
-  static std::unordered_set<uint32_t> enabledExtensions;
+  static std::unordered_set<uint64_t> enabledExtensions;
 
   static VkBool32 debugCallback(VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity, VkDebugUtilsMessageTypeFlagsEXT messageTypes, const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData, void* pUserData);
   
   static void create(std::vector<const char*> extensions);
   /**
-   * Check for availability of instance extensions. This function <em>must</em> be called before using any functionality provided by an instance extension.
-   * <c>if (GraphicsInstance::extensionEnabled(Tools::hash(VK_EXT_DEBUG_UTILS_EXTENSION_NAME))) { // code that names objects }</c>
+   * Check the availability of instance extensions. This function <em>must</em> be called before using any functionality provided by an instance extension.
+   *
+   * <c>if (GraphicsInstance::extensionEnabled(Tools::hash(VK_EXT_DEBUG_UTILS_EXTENSION_NAME))) { //code that names objects }</c>
    * @param nameHash The hash of the extension's name.
    * @return <c>true</c> if the extension is enabled and <c>false</c> if not.
    */
